# cms-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Go backend migration platform API.",
        "title": "CMS API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "登录",
                "parameters": [
                    {
                        "description": "登录请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.loginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.loginResponseEnvelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/auth.errorResponseEnvelope"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/auth.errorResponseEnvelope"
                        },
                        "headers": {
                            "Retry-After": {
                                "type": "string",
                                "description": "等待重试的秒数"
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "认证"
                ],
                "summary": "退出登录",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.emptyResponseEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/auth.errorResponseEnvelope"
                        }
                    }
                }
            }
        },
        "/api/auth/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "认证"
                ],
                "summary": "当前用户信息",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.currentUserResponseEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/auth.errorResponseEnvelope"
                        }
                    }
                }
            }
        },
        "/api/auth/menus": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "认证"
                ],
                "summary": "当前用户可见菜单",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.currentUserMenusResponseEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/auth.errorResponseEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "auth.LoginResult": {
            "type": "object",
            "properties": {
                "expiresIn": {
                    "type": "integer"
                },
                "tokenName": {
                    "type": "string"
                },
                "tokenValue": {
                    "type": "string"
                }
            }
        },
        "auth.currentUserDeptResponse": {
            "type": "object",
            "properties": {
                "deptCode": {
                    "type": "string"
                },
                "deptName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "auth.currentUserMenuResponse": {
            "type": "object",
            "properties": {
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/auth.currentUserMenuResponse"
                    }
                },
                "component": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "menuName": {
                    "type": "string"
                },
                "menuType": {
                    "type": "string"
                },
                "parentId": {
                    "type": "integer"
                },
                "path": {
                    "type": "string"
                },
                "permissionCode": {
                    "type": "string"
                },
                "sortOrder": {
                    "type": "integer"
                },
                "visible": {
                    "type": "integer"
                }
            }
        },
        "auth.currentUserMenusResponseEnvelope": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/auth.currentUserMenuResponse"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "auth.currentUserResponse": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "dept": {
                    "$ref": "#/definitions/auth.currentUserDeptResponse"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastLoginIp": {
                    "type": "string"
                },
                "lastLoginTime": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/auth.currentUserRoleResponse"
                    }
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "auth.currentUserResponseEnvelope": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/auth.currentUserResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "auth.currentUserRoleResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "roleCode": {
                    "type": "string"
                },
                "roleName": {
                    "type": "string"
                }
            }
        },
        "auth.emptyResponseEnvelope": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "auth.errorResponseEnvelope": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "auth.loginRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "auth.loginResponseEnvelope": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/auth.LoginResult"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}